remove deploy& send to under contract
